apt-requirements-bionic.txt
clang-8
clang-format-8
clang-tidy-8
cmake (>= 3.5.0)
debhelper (>= 9)
doxygen
libargue-dev
libclang-8-dev
libfmt-dev
libgoogle-glog-dev
libprotobuf-dev
libprotoc-dev
libpython3-dev
libtangent-util-dev
llvm-8-dev
ninja-build
pkg-config
protobuf-compiler
python
python3
python3-dev
python3-protobuf
python3-jinja2
quilt

